Reworking docs and internal naming of stuff
